requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
ccxt==4.5.47
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
coincurve==21.0.0
colorama==0.4.6
cryptography==46.0.6
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.151.11
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sortedcontainers==2.4.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
yarl==1.23.0
